WIP Complete refactor
