+++ home/estasney/mysite/templates/keywords.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %}
 {% block content %}
 {% if success == 'True' %}
-<form action="." method="POST">
+<form action="/keywords" method="POST">
 <div class="form-group">
   <label for="raw_text">Enter Text to Find Keywords</label>
   <textarea class="form-control" rows="5" id="raw_text" name="raw_text">{{ original }}</textarea>
@@ -11,7 +11,7 @@
 </div>
 </form>
 {% else %}
-<form action="." method="POST">
+<form action="/keywords" method="POST">
 <div class="form-group">
   <label for="raw_text">Enter Text to Find Keywords</label>
   <textarea class="form-control" rows="5" id="raw_text" name="raw_text"></textarea>
