+++ home/estasney/mysite/templates/related.html
@@ -1,7 +1,7 @@
 {% extends "layout.html" %}
 {% block content %}
 {% if success == 'True' %}
- <form action="." method="POST">
+ <form action="/related" method="POST">
     <div class="row">
       <div class="col-lg-4 col-lg-offset-4">
       <div class="input-group">
@@ -15,7 +15,7 @@
     </div>
   </form>
 {% else %}
- <form action="." method="POST">
+ <form action="/related" method="POST">
     <div class="row">
       <div class="col-lg-4 col-lg-offset-4">
       <div class="input-group">
